Handle Nursery Request listing. First draft implementation.

+++ app/models/ExampleModel.class.php
@@ -90,7 +90,7 @@ class ExampleModel{
     /**
      * Get the value of id
      */ 
-    public function getId():long{
+    public function getId():int{
         return $this->id;
     }
 
@@ -98,7 +98,7 @@ class ExampleModel{
      * Set the value of id
      *
      */ 
-    public function setId(long $id){
+    public function setId(int $id){
         $this->id = $id;
     }
